+++ CHANGELOG.md
@@ -6,3 +6,5 @@ All notable changes to `NovaPoshta` will be documented in this file
 - [Add][ND] `getCatalogCounterparty` in Counterparty
 - [Add] `getSettlements` in Address
 - [Add][ND] `prolongateKey` in CommonGeneral (NOT WORK)
+- [Add] `delete` in InternetDocument
+- [Fix] `checkTTN` in TrackingDocument